Changes for upload

+++ app/Http/Controllers/BestDealController.php
@@ -20,8 +20,12 @@ class BestDealController extends Controller
      */
     public function index()
     {
-        $best_deals = BestDeal::latest()->get();
-        return view('backend.deals.index', compact('best_deals'));
+        if (auth()->user()->can('View Deals')) {
+            $best_deals = BestDeal::latest('id')->get();
+            return view('backend.deals.index', compact('best_deals'));
+        } else {
+            abort('404');
+        }
     }
 
     /**
@@ -31,8 +35,12 @@ class BestDealController extends Controller
      */
     public function create()
     {
-        $products = Product::where('status', 1)->select('id', 'title')->get();
-        return view('backend.deals.create', compact('products'));
+        if (auth()->user()->can('View Deals')) {
+            $products = Product::where('status', 1)->select('id', 'title')->latest('id')->get();
+            return view('backend.deals.create', compact('products'));
+        } else {
+            abort('404');
+        }
     }
 
     /**
@@ -43,51 +51,54 @@ class BestDealController extends Controller
      */
     public function store(Request $request)
     {
-        // return $request;
-        $request->validate([
-            'date' => ['date', 'required',],
-            'time' => [ 'required',],
-            'discount' => ['required', 'numeric', 'min:2', 'max:99'],
-            'title' => ['required',],
-            'product_id.*' => ['required'],
-        ]);
-        if ($request->product_id != '') {
-            $deal = new BestDeal;
-            $deal->title = $request->title;
-            $deal->discount = $request->discount;
-            $deal->expire_date = $request->date;
-            $deal->expire_time = $request->time;
-            if ($request->deal_banner != '') {
-                if ($request->hasFile('deal_banner')) {
-                    $deal_banner = $request->file('deal_banner');
-                    $extension = config('app.name') . '-' . Str::random(2) . '.' . $deal_banner->getClientOriginalExtension();
-                    Image::make($deal_banner)->save(public_path('deal_banner/' . $extension), 90);
-                    $deal->deal_banner = $extension;
+        if (auth()->user()->can('View Deals')) {
+            $request->validate([
+                'date' => ['date', 'required',],
+                'time' => ['required',],
+                'discount' => ['required', 'numeric', 'min:2', 'max:99'],
+                'title' => ['required',],
+                'product_id.*' => ['required'],
+            ]);
+            if ($request->product_id != '') {
+                $deal = new BestDeal;
+                $deal->title = $request->title;
+                $deal->discount = $request->discount;
+                $deal->expire_date = $request->date;
+                $deal->expire_time = $request->time;
+                if ($request->deal_banner != '') {
+                    if ($request->hasFile('deal_banner')) {
+                        $deal_banner = $request->file('deal_banner');
+                        $extension = config('app.name') . '-' . Str::random(2) . '.' . $deal_banner->getClientOriginalExtension();
+                        Image::make($deal_banner)->save(public_path('deal_banner/' . $extension), 90);
+                        $deal->deal_banner = $extension;
+                    }
+                } else {
+                    $request->validate(['deal_backgraound_color' => ['required']]);
+                    $deal->deal_backgraound_color = $request->deal_backgraound_color;
                 }
-            } else {
-                $request->validate(['deal_backgraound_color' => ['required']]);
-                $deal->deal_backgraound_color = $request->deal_backgraound_color;
-            }
-            $deal->save();
+                $deal->save();
 
-            foreach ($request->product_id as $key => $product_id) {
+                foreach ($request->product_id as $key => $product_id) {
 
-                $attributes = Attribute::where('product_id', $product_id)->get();
-                foreach ($attributes as $key => $attribute) {
-                    $attribute->discount = $request->discount;
-                    $discount_amount = ($attribute->regular_price * $request->discount) / 100;
-                    $sell_price = round($attribute->regular_price - $discount_amount);
-                    $attribute->sell_price = $sell_price;
-                    $attribute->save();
+                    $attributes = Attribute::where('product_id', $product_id)->get();
+                    foreach ($attributes as $key => $attribute) {
+                        $attribute->discount = $request->discount;
+                        $discount_amount = ($attribute->regular_price * $request->discount) / 100;
+                        $sell_price = round($attribute->regular_price - $discount_amount);
+                        $attribute->sell_price = $sell_price;
+                        $attribute->save();
+                    }
+                    $deal_product = new BestDealProduct;
+                    $deal_product->best_deal_id = $deal->id;
+                    $deal_product->product_id = $product_id;
+                    $deal_product->save();
                 }
-                $deal_product = new BestDealProduct;
-                $deal_product->best_deal_id = $deal->id;
-                $deal_product->product_id = $product_id;
-                $deal_product->save();
-            }
-            return redirect()->route('deals.index')->with('success', 'Added Successfully');
-        };
-        return back();
+                return redirect()->route('deals.index')->with('success', 'Added Successfully');
+            };
+            return back();
+        } else {
+            abort('404');
+        }
     }
 
     /**
@@ -98,39 +109,43 @@ class BestDealController extends Controller
      */
     public function show($id)
     {
-        $BestDeal =  BestDeal::findorfail($id);
-        if ($BestDeal->status == 1) {
-            $best_deal_product = BestDealProduct::where('best_deal_id', $BestDeal->id)->get();
-            foreach ($best_deal_product as $key => $deal_product) {
-                $attributes = Attribute::where('product_id', $deal_product->product_id)->get();
-                foreach ($attributes as $key => $attribute) {
-                    $attribute->discount = '';
-                    $attribute->sell_price = '';
-                    $attribute->save();
+        if (auth()->user()->can('View Deals')) {
+            $BestDeal =  BestDeal::findorfail($id);
+            if ($BestDeal->status == 1) {
+                $best_deal_product = BestDealProduct::where('best_deal_id', $BestDeal->id)->get();
+                foreach ($best_deal_product as $key => $deal_product) {
+                    $attributes = Attribute::where('product_id', $deal_product->product_id)->get();
+                    foreach ($attributes as $key => $attribute) {
+                        $attribute->discount = '';
+                        $attribute->sell_price = '';
+                        $attribute->save();
+                    }
                 }
-            }
-            $BestDeal->status = 2;
-            $BestDeal->save();
-            return back()->with('warning', 'Inactive Successfully');
-        } else {
+                $BestDeal->status = 2;
+                $BestDeal->save();
+                return back()->with('warning', 'Inactive Successfully');
+            } else {
 
-            if (Carbon::today()->format('Y-m-d') > $BestDeal->expire_date) {
-                return back()->with('warning', 'Out of Date');;
-            }
-            $best_deal_product = BestDealProduct::where('best_deal_id', $BestDeal->id)->get();
-            foreach ($best_deal_product as $key => $deal_product) {
-                $attributes = Attribute::where('product_id', $deal_product->product_id)->get();
-                foreach ($attributes as $key => $attribute) {
-                    $attribute->discount = $BestDeal->discount;
-                    $discount_amount = ($attribute->regular_price * $BestDeal->discount) / 100;
-                    $sell_price = round($attribute->regular_price - $discount_amount);
-                    $attribute->sell_price = $sell_price;
-                    $attribute->save();
+                if (Carbon::today()->format('Y-m-d') > $BestDeal->expire_date) {
+                    return back()->with('warning', 'Out of Date');;
                 }
+                $best_deal_product = BestDealProduct::where('best_deal_id', $BestDeal->id)->get();
+                foreach ($best_deal_product as $key => $deal_product) {
+                    $attributes = Attribute::where('product_id', $deal_product->product_id)->get();
+                    foreach ($attributes as $key => $attribute) {
+                        $attribute->discount = $BestDeal->discount;
+                        $discount_amount = ($attribute->regular_price * $BestDeal->discount) / 100;
+                        $sell_price = round($attribute->regular_price - $discount_amount);
+                        $attribute->sell_price = $sell_price;
+                        $attribute->save();
+                    }
+                }
+                $BestDeal->status = 1;
+                $BestDeal->save();
+                return back()->with('success', 'Active Successfully');
             }
-            $BestDeal->status = 1;
-            $BestDeal->save();
-            return back()->with('success', 'Active Successfully');
+        } else {
+            abort('404');
         }
     }
 
@@ -163,24 +178,28 @@ class BestDealController extends Controller
      */
     public function destroy($id)
     {
-        $BestDeal =  BestDeal::findorfail($id);
-        if ($BestDeal->deal_banner) {
-            $old_banner = public_path('deal_banner/' . $BestDeal->deal_banner);
-            if (file_exists($old_banner)) {
-                unlink($old_banner);
+        if (auth()->user()->can('View Deals')) {
+            $BestDeal =  BestDeal::findorfail($id);
+            if ($BestDeal->deal_banner) {
+                $old_banner = public_path('deal_banner/' . $BestDeal->deal_banner);
+                if (file_exists($old_banner)) {
+                    unlink($old_banner);
+                }
             }
-        }
-        $best_deal_product = BestDealProduct::where('best_deal_id', $BestDeal->id)->get();
-        foreach ($best_deal_product as $key => $deal_product) {
-            $attributes = Attribute::where('product_id', $deal_product->product_id)->get();
-            foreach ($attributes as $key => $attribute) {
-                $attribute->discount = '';
-                $attribute->sell_price = '';
-                $attribute->save();
+            $best_deal_product = BestDealProduct::where('best_deal_id', $BestDeal->id)->get();
+            foreach ($best_deal_product as $key => $deal_product) {
+                $attributes = Attribute::where('product_id', $deal_product->product_id)->get();
+                foreach ($attributes as $key => $attribute) {
+                    $attribute->discount = '';
+                    $attribute->sell_price = '';
+                    $attribute->save();
+                }
+                $deal_product->delete();
             }
-            $deal_product->delete();
+            $BestDeal->delete();
+            return back()->with('delete', 'Deleted Successfully');
+        } else {
+            abort('404');
         }
-        $BestDeal->delete();
-        return back()->with('delete', 'Deleted Successfully');
     }
 }
